requirements.txt
aiohttp
fake_useragent
pyyaml
colorama